LPS-61560 Extract the image selector taglib
