+++ modules/frontend/frontend-taglib/src/main/resources/META-INF/resources/image_selector/init.jsp
@@ -14,6 +14,6 @@
  */
 --%>
 
-<%@ include file="/html/taglib/init.jsp" %>
+<%@ include file="/init.jsp" %>
 
 <portlet:defineObjects />
