+++ modules/frontend/frontend-taglib/src/main/resources/META-INF/resources/image_selector/js/image_selector.js
