+++ portal-web/docroot/html/common/init.jsp
@@ -235,8 +235,7 @@ page import="java.util.Set" %><%@
 page import="java.util.TimeZone" %><%@
 page import="java.util.TreeSet" %>
 
-<%@ page import="javax.portlet.ActionRequest" %><%@
-page import="javax.portlet.MimeResponse" %><%@
+<%@ page import="javax.portlet.MimeResponse" %><%@
 page import="javax.portlet.PortletConfig" %><%@
 page import="javax.portlet.PortletContext" %><%@
 page import="javax.portlet.PortletException" %><%@
